1. moved runningGenerateRiverFlows function into generate_river_flows.py. 2. renamed function 'prepareGEFSdata' to 'prepareWeatherForecastData.'
